techpack: audio: Fix uninitialized header usage
When the cmi header is filled, it isn't zeroed out; it's updated, so the
previous value it contains is retained as part of a bitmask. This
constitutes uninitialized usage of the cmi header struct.

Fix it by zeroing out the header with a memset, as is done elsewhere in
this driver.

This fixes the following warnings:
sound/soc/codecs/wcd_cpe_core.c: In function 'wcd_cpe_lab_ch_setup':
include/sound/cpe_cmi.h:173:18: error: 'lab_eob.hdr_info' may be used uninitialized in this function [-Werror=maybe-uninitialized]

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
Signed-off-by: Adam W. Willis <return.of.octobot@gmail.com>
Change-Id: I46c94fd39b551ec08152fc5a7ec3bffb9f4e4b0a
Signed-off-by: Diaz1401 <diaznura@outlook.com>